Created New Java File Menu.java and Updates to App.java for Code Readability
